Model example that can be used to test other functions of VarDecomp. A model with a covariance matrix.

Model example that can be used to test other functions of VarDecomp. A model with a covariance matrix.

data(mod_cov, package = "VarDecomp")

Format

An object of class brmsfit with the following components:

formula

The formula or formulas specifying the model.

data

The data frame containing the variables used in the model.

prior

The prior distributions for the model parameters.

sample_prior

The prior samples, if sample_prior was set to TRUE during fitting.

stanmodel

The Stan model code used for fitting.

stanfit

The fitted model object returned by Stan.

algorithm

The algorithm used for fitting the model, typically 'sampling' or 'meanfield'.

family

The family object specifying the model family (e.g., Gaussian, Bernoulli).

data2

Additional data used in the model, if any.

ranef

A list of random effects structures in the model.

autocor

The autocorrelation structure, if any, used in the model.

cov_ranef

The covariance matrix of the random effects.

smooth

A list of smooth terms in the model, if any.

stanvars

A list of Stan variables defined in the model code.

inits

Initial values used for the parameters in the Stan model.

fun

The function used for sampling or variational inference.

stan_args

A list of arguments passed to the Stan function.

control

A list of control parameters for the Stan sampling algorithm.

mcmc_levs

A list of MCMC settings used in the model fitting.

converged

A logical indicating if the model has converged.

loo

The leave-one-out cross-validation (LOO) results for the model, if computed.

waic

The Watanabe-Akaike Information Criterion (WAIC) results for the model, if computed.

log_lik

The log-likelihood values for the model, if computed.

fitted

The fitted values from the model.

residuals

The residuals from the model.

summary

A summary of the model, including parameter estimates and diagnostics.

posterior_samples

A matrix of posterior samples for the model parameters.

diagnostics

Diagnostic information for the model, including convergence diagnostics.

user_data

Any additional user data stored in the model object.

Source

Generated using the brms_model function.

See also